refactor(signer): 'SignaturePublisherDmq' uses a 'DmqPublisher'
