+++ mithril-signer/src/services/signature_publisher/mod.rs
@@ -1,12 +1,12 @@
 mod delayer;
+mod dmq;
 mod http;
 mod interface;
 mod noop;
-mod pallas;
 mod retrier;
 
 pub use delayer::*;
+pub use dmq::*;
 pub use interface::*;
 pub use noop::*;
-pub use pallas::*;
 pub use retrier::*;
